resources, requests

+++ app/Http/Controllers/BookController.php
@@ -5,7 +5,13 @@ namespace App\Http\Controllers;
 use App\Models\Book;
 use Illuminate\Http\Request;
 
-use App\Http\Requests\BookRequest;
+// BookRequests
+use App\Http\Requests\Api\v1\BookRequests\BookStoreRequest;
+use App\Http\Requests\Api\v1\BookRequests\BookUpdateRequest;
+
+// BookResources
+use App\Http\Resources\Api\v1\BookResources\BookResource;
+use App\Http\Resources\Api\v1\BookResources\BookWithAuthorNameResource;
 
 class BookController extends Controller
 {
@@ -22,7 +28,7 @@ class BookController extends Controller
             data: [
                 'error' => 0,
                 'message' => __('Список книг успешно получен.'),
-                'books' => $books,
+                'books' => BookResource::collection($books),
             ],
             status: 200,
         );
@@ -30,21 +36,13 @@ class BookController extends Controller
 
     public function booksWithAuthorName()
     {
-        $books = Book::select(
-            'books.id',
-            'books.name',
-            'books.genre',
-            'books.year',
-            'authors.name as authorName',
-        )
-            ->leftJoin('authors', 'books.author_id', '=', 'authors.id')
-            ->get();
+        $books = Book::all();
 
         return response()->json(
             data: [
                 'error' => 0,
                 'message' => __('Список книг с именами авторов успешно получен.'),
-                'books' => $books,
+                'books' => BookWithAuthorNameResource::collection($books),
             ],
             status: 200,
         );
@@ -56,10 +54,11 @@ class BookController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(BookRequest $request)
+    public function store(BookStoreRequest $request)
     {
         $validated = $request->validated();
 
+        // TODO: check if admin
         if (!isset($validated['author_id'])) {
             $validated['author_id'] = (int) $request->user()->id;
         }
@@ -71,7 +70,7 @@ class BookController extends Controller
             data: [
                 'error' => 0,
                 'message' => __('Книга успешно добавлена.'),
-                'book' => $book,
+                'book' => new BookResource($book),
             ],
             status: 200,
         );
@@ -92,7 +91,7 @@ class BookController extends Controller
                 data: [
                     'error' => 0,
                     'message' => __('Книга успешно получена.'),
-                    'book' => $book,
+                    'book' => new BookResource($book),
                 ],
                 status: 200,
             );
@@ -114,9 +113,32 @@ class BookController extends Controller
      * @param  \App\Models\Book  $book
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(BookUpdateRequest $request, $id)
     {
-        //
+        $book = Book::find($id);
+        $validated = $request->validated();
+        
+        if ($book) {
+            $this->authorize('update', [Book::class, $book]);
+
+            $book->update($validated);
+            return response()->json(
+                data: [
+                    'error' => 0,
+                    'message' => __('Книга успешно обновлена.'),
+                    'book' => new BookResource($book),
+                ],
+                status: 200,
+            );
+        } else {
+            return response()->json(
+                data: [
+                    'error' => 1,
+                    'message' => __('Книга не найдена.'),
+                ],
+                status: 404,
+            );
+        }
     }
 
     /**
@@ -127,15 +149,17 @@ class BookController extends Controller
      */
     public function destroy($id)
     {
-        $book = Book::find($id);
+        $book = Book::where('id', $id)->first();
 
         if ($book) {
+            $this->authorize('delete', [Book::class, $book]);
+
             $book->delete();
             return response()->json(
                 data: [
                     'error' => 0,
                     'message' => __('Книга успешно удалена.'),
-                    'book' => $book,
+                    'book' => new BookResource($book),
                 ],
                 status: 200,
             );
